move plug to its own module
